Strategy, IRs, Sonars, Website... - Lots of additions to strategy - Start of strategy tuning with tests - IRs work via ADC on FPGA - Sonar filtering corrected - Website: complete redesign of charts page, adding a lot of   information.

+++ MyApp_v2/RobinsunFiles/ctrl_io.h
@@ -125,7 +125,7 @@ typedef struct CtrlIn
     double odo_r_speed;
     double odo_l_speed;
     
-    double sonars[6];
+    int sonars[6];
     
     double speed_blocks;
     double speed_fish_horizontal;
